requirements.txt
pillow>=9.0.0
numpy>=1.21.0
scipy>=1.7.0
torch>=1.9.0
opencv-python>=4.5.0
pywavelets>=1.1.0
pillow-heif>=0.15.0
rawpy>=0.19.0
imageio>=2.31.0